move `fieldRef()` to `model.ts`
